+++ products/statement-generator/src/App.tsx
@@ -36,8 +36,7 @@ import NotFound from 'pages/NotFound';
 import PreviewPage from 'pages/PreviewPage';
 import PrivacyPolicy from 'pages/PrivacyPolicy';
 import TermsOfUse from 'pages/TermsOfUse';
-// import FAQ from 'pages/FAQ';
-import FAQ2 from 'pages/FAQ2';
+import FAQ from 'pages/FAQ';
 import AboutUs from 'pages/AboutUs';
 
 import 'styles/App.css';
@@ -177,7 +176,7 @@ const App: React.FC = () => {
                     component={PrivacyPolicy}
                   />
                   <Route path={AppUrl.TermsOfUse} component={TermsOfUse} />
-                  <Route path={AppUrl.FAQ} component={FAQ2} />
+                  <Route path={AppUrl.FAQ} component={FAQ} />
                   <Route path={AppUrl.AboutUs} component={AboutUs} />
 
                   <Route component={NotFound} />
