+++ products/statement-generator/src/pages/FAQ.tsx
@@ -1,103 +1,251 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-
-import LinkAsText from 'components/LinkAsText';
-
 import ContentContainer from 'components-layout/ContentContainer';
+import { makeStyles, createStyles } from '@material-ui/core';
+import { FAQAccordion } from 'components/FAQAccordion';
+import faqDesktop from '../assets/faqDesktop.svg';
+import faqMobile from '../assets/faqMobile.svg';
 
-function FAQ() {
+interface IContentContainer {
+  children?: any;
+  className?: string;
+}
+
+const useStyles = makeStyles(({ palette, breakpoints, spacing }) =>
+  createStyles({
+    Header: {
+      width: '100vw',
+      display: 'flex',
+      justifyContent: 'center',
+      backgroundColor: palette.primary.lighter,
+      marginBottom: '12px',
+
+      [breakpoints.down('sm')]: {
+        minHeight: '509px',
+      },
+    },
+    HeaderContent: {
+      maxWidth: '996px',
+      minWidth: '342px',
+      width: '69.2%',
+      margin: '45px 222px 54.64px',
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+
+      [breakpoints.down('md')]: {
+        flexDirection: 'column',
+        width: '100%',
+        margin: '24px 16px',
+        gap: '36px',
+      },
+    },
+    HeaderText: {
+      display: 'flex',
+      flexDirection: 'column',
+      justifyContent: 'center',
+      alignItems: 'center',
+      backgroundColor: palette.primary.lighter,
+      marginRight: '113.5px',
+
+      [breakpoints.down('md')]: {
+        marginRight: 0,
+      },
+
+      [breakpoints.down('sm')]: {
+        gap: '24px',
+      },
+    },
+    Heading: {
+      fontSize: '34px',
+      lineHeight: '51px',
+      marginBotton: '16px',
+      fontWeight: 700,
+    },
+    SubHeading: {
+      fontSize: '24px',
+      lineHeight: '36px',
+      fontWeight: 400,
+    },
+    ImgContainer: {
+      width: '100%',
+      textAlign: 'center',
+    },
+    Img: {
+      width: 'auto',
+      maxWidth: '100%',
+      minWidth: '182px',
+    },
+    FAQContainer: {
+      maxWidth: '996px',
+      minWidth: '300px',
+      width: '69.2%',
+      padding: spacing(3),
+      paddingTop: 0,
+      marginLeft: 'auto',
+      marginRight: 'auto',
+      display: 'flex',
+      flex: '1 0 auto',
+      flexDirection: 'column',
+
+      [breakpoints.down('sm')]: {
+        padding: spacing(2),
+      },
+
+      [breakpoints.down('xs')]: {
+        marginLeft: 'initial',
+        marginRight: 'initial',
+        width: '100%',
+      },
+    },
+    sectionTitle: {
+      color: palette.primary.main,
+      fontSize: '24px',
+      lineHeight: '36px',
+      fontWeight: 600,
+      boxShadow: 'none',
+      border: 'none',
+      marginTop: '24px',
+      marginBottom: '12px',
+
+      [breakpoints.down('sm')]: {
+        marginTop: '24px',
+      },
+    },
+  })
+);
+
+export const FAQHeader: React.FC = () => {
+  const classes = useStyles();
   const { t } = useTranslation();
 
   return (
-    <ContentContainer>
-      <h2>{t('faq_page.page_header')}</h2>
+    <header className={classes.Header}>
+      <div className={classes.HeaderContent}>
+        <div className={classes.HeaderText}>
+          <h2 className={classes.Heading}>{t('faq_page.page_header')}</h2>
+          <p className={classes.SubHeading}>{t('faq_page.page_subheader')}</p>
+        </div>
+        <div className={classes.ImgContainer}>
+          <img
+            className={classes.Img}
+            src={faqDesktop}
+            alt="Two people talking"
+          />
+        </div>
+      </div>
+    </header>
+  );
+};
 
-      <h3>{t('faq_page.title1')}</h3>
-      <p>{t('faq_page.paragraph1')}</p>
+const FAQContainer = ({ children, className = '' }: IContentContainer) => {
+  const classes = useStyles();
 
-      <h3>{t('faq_page.title2')}</h3>
-      <p>{t('faq_page.paragraph2')}</p>
-      <p>{t('faq_page.paragraph2b')}</p>
-      <p>{t('faq_page.paragraph2c')}</p>
-      <p>{t('faq_page.paragraph2d')}</p>
-      <p>{t('faq_page.paragraph2e')}</p>
+  return (
+    <div className={`${classes.FAQContainer} ${className}`}>{children}</div>
+  );
+};
 
-      <h3>{t('faq_page.title3')}</h3>
-      <p>{t('faq_page.paragraph3')}</p>
-      <p>{t('faq_page.paragraph3b')}</p>
+function FAQ() {
+  const classes = useStyles();
+  const { t } = useTranslation();
 
-      <h3>{t('faq_page.title4')}</h3>
-      <p>{t('faq_page.paragraph4')}</p>
-      <p>{t('faq_page.paragraph4b')}</p>
-      <p>{t('faq_page.paragraph4c')}</p>
+  return (
+    <>
+      <FAQHeader />
+      <FAQContainer>
+        <section>
+          <h3 className={classes.sectionTitle}>{t('faq_page.section1')}</h3>
+          <FAQAccordion
+            title="faq_page.title1.1"
+            content={[
+              'faq_page.paragraph1.1a',
+              'faq_page.paragraph1.1b',
+              'faq_page.paragraph1.1c',
+              'faq_page.link-hfla',
+            ]}
+          />
+          <FAQAccordion
+            title="faq_page.title1.2"
+            content={['faq_page.paragraph1.2a', 'faq_page.paragraph1.2b']}
+          />
+          <FAQAccordion
+            title="faq_page.title1.3"
+            content={['faq_page.paragraph1.3a', 'faq_page.paragraph1.3b']}
+          />
+          <FAQAccordion
+            title="faq_page.title1.4"
+            content={[
+              'faq_page.paragraph1.4a',
+              'faq_page.paragraph1.4b',
+              'faq_page.paragraph1.4c',
+              'faq_page.paragraph1.4d',
+            ]}
+          />
+        </section>
 
-      <h3>{t('faq_page.title5')}</h3>
-      <p>{t('faq_page.paragraph5')}</p>
-      <p>{t('faq_page.paragraph5b')}</p>
+        <section>
+          <h2 className={classes.sectionTitle}>{t('faq_page.section2')}</h2>
+          <FAQAccordion
+            title="faq_page.title2.1"
+            content={['faq_page.paragraph2.1a', 'faq_page.paragraph2.1b']}
+          />
+          <FAQAccordion
+            title="faq_page.title2.2"
+            content={['faq_page.paragraph2.2a', 'faq_page.paragraph2.2b']}
+          />
+          <FAQAccordion
+            title="faq_page.title2.3"
+            content={['faq_page.paragraph2.3a', 'faq_page.paragraph2.3b']}
+          />
+          <FAQAccordion
+            title="faq_page.title2.4"
+            content={['faq_page.paragraph2.4a', 'faq_page.paragraph2.4b']}
+          />
+          <FAQAccordion
+            title="faq_page.title2.5"
+            content={['faq_page.paragraph2.5a', 'faq_page.paragraph2.5b']}
+          />
+        </section>
 
-      <h3>{t('faq_page.title6')}</h3>
-      <p>{t('faq_page.paragraph6')}</p>
-      <p>{t('faq_page.paragraph6b')}</p>
-      <ul>
-        <li>{t('faq_page.paragraph6b_li1')}</li>
-        <li>{t('faq_page.paragraph6b_li2')}</li>
-        <li>{t('faq_page.paragraph6b_li3')}</li>
-      </ul>
-      <p>{t('faq_page.paragraph6c')}</p>
-      <p>{t('faq_page.paragraph6d')}</p>
-      <ul>
-        <li>{t('faq_page.paragraph6d_li1')}</li>
-        <li>{t('faq_page.paragraph6d_li2')}</li>
-        <li>{t('faq_page.paragraph6d_li3')}</li>
-        <li>{t('faq_page.paragraph6d_li4')}</li>
-        <li>{t('faq_page.paragraph6d_li5')}</li>
-        <li>{t('faq_page.paragraph6d_li6')}</li>
-      </ul>
-      <p>
-        {t('faq_page.paragraph6e')}
-        <LinkAsText link="https://www.courts.ca.gov/documents/Prop47FAQs.pdf" />
-      </p>
-
-      <h3>{t('faq_page.title7')}</h3>
-      <p>{t('faq_page.paragraph7')}</p>
-      <p>{t('faq_page.paragraph7b')}</p>
-      <ul>
-        <li>{t('faq_page.paragraph7b_li1')}</li>
-        <li>{t('faq_page.paragraph7b_li2')}</li>
-        <li>{t('faq_page.paragraph7b_li3')}</li>
-        <li>{t('faq_page.paragraph7b_li4')}</li>
-      </ul>
-
-      <p>{t('faq_page.paragraph7c')}</p>
-      <p>{t('faq_page.paragraph7d')}</p>
-      <ul>
-        <li>{t('faq_page.paragraph7d_li1')}</li>
-        <li>{t('faq_page.paragraph7d_li2')}</li>
-        <li>{t('faq_page.paragraph7d_li3')}</li>
-        <li>{t('faq_page.paragraph7d_li4')}</li>
-      </ul>
-      <p>{t('faq_page.paragraph7e')}</p>
-      <p>{t('faq_page.paragraph7f')}</p>
-      <p>
-        {t('faq_page.paragraph7g')}
-        <LinkAsText link="https://drugpolicy.org/sites/default/files/Prop64-Resentencing-Guide-July2017.pdf" />
-      </p>
-
-      <h3>{t('faq_page.title8')}</h3>
-      <p>{t('faq_page.paragraph8')}</p>
-      <p>{t('faq_page.paragraph8b')}</p>
-
-      <h3>{t('faq_page.title9')}</h3>
-      <p>{t('faq_page.paragraph9')}</p>
-
-      <h3>{t('faq_page.title10')}</h3>
-      <p>{t('faq_page.paragraph10')}</p>
-      <p>{t('faq_page.paragraph10b')}</p>
-
-      <h3>{t('faq_page.title11')}</h3>
-      <p>{t('faq_page.paragraph11')}</p>
-      <p>{t('faq_page.paragraph11b')}</p>
-    </ContentContainer>
+        <section>
+          <h2 className={classes.sectionTitle}>{t('faq_page.section3')}</h2>
+          <FAQAccordion
+            title="faq_page.title3.1"
+            content={[
+              'faq_page.paragraph3.1a',
+              'faq_page.paragraph3.1b',
+              'faq_page.paragraph3.1c',
+              'faq_page.paragraph3.1d',
+            ]}
+          />
+          <FAQAccordion
+            title="faq_page.title3.2"
+            content={[
+              'faq_page.paragraph3.2a',
+              'faq_page.paragraph3.2b',
+              'faq_page.paragraph3.2c',
+            ]}
+          />
+          <FAQAccordion
+            title="faq_page.title3.3"
+            content={[
+              'faq_page.paragraph3.3a',
+              'faq_page.paragraph3.3b',
+              'faq_page.paragraph3.3c',
+            ]}
+          />
+          <FAQAccordion
+            title="faq_page.title3.4"
+            content={['faq_page.paragraph3.4a', 'faq_page.paragraph3.4b']}
+          />
+          <FAQAccordion
+            title="faq_page.title3.5"
+            content={['faq_page.paragraph3.5a']}
+          />
+        </section>
+      </FAQContainer>
+    </>
   );
 }
